move exception handling to ControllerAdvice
This allows us to have global exception handling without needing
to extend the same base controller
